K5.c now working (was trouble in \n in first line)

+++ Ks/K5.c
@@ -1,12 +1,9 @@
-#include <sys/stat.h>
-#include <sys/types.h>
 #include <unistd.h>
 #include <fcntl.h>
-#include <wait.h>
 #include <stdlib.h>
 #include <string.h>
 
-enum { BUF_SIZE = 10 };
+enum { BUF_SIZE = 1024 };
 
 int 
 main (int argc, char **argv) {
@@ -18,11 +15,15 @@ main (int argc, char **argv) {
     char *file_name = *argv;
     int fd = open(file_name, O_RDONLY);
 
+    if (fd == -1) {
+        return 0;
+    }
+
     char temp_name[7] = "XXXXXX";
     int temp_fd = mkstemp(temp_name);
     unlink(temp_name);
 
-    char *buf = calloc(BUF_SIZE + 1, 1);
+    static char buf[BUF_SIZE + 1];
 
     int cnt_readed_byte = 0;
     char *pos_end_first_str;
@@ -39,14 +40,14 @@ main (int argc, char **argv) {
     }
     
     //Write end of first str
-    write(temp_fd, buf, pos_end_first_str - buf);
+    write(temp_fd, buf, pos_end_first_str - buf + 1);
 
     //Find end of second str
     char *pos_end_second_str = strchr(pos_end_first_str + 1, '\n');
 
     if (pos_end_second_str != NULL) {
         //Write file after second str
-        write(temp_fd, pos_end_second_str, (buf + strlen(buf)) - pos_end_second_str);
+        write(temp_fd, pos_end_second_str + 1, (buf + strlen(buf)) - pos_end_second_str - 1);
     } else {
         //Skipping second str
         do {
@@ -55,7 +56,7 @@ main (int argc, char **argv) {
         } while (!(((pos_end_second_str = strchr(buf, '\n')) != NULL) || (cnt_readed_byte == 0)));
 
         //Write file after second str
-        write(temp_fd, pos_end_second_str, (buf + strlen(buf)) - pos_end_second_str);
+        write(temp_fd, pos_end_second_str + 1, (buf + strlen(buf)) - pos_end_second_str -1);
     }
 
     //Copy file after second str
@@ -80,6 +81,5 @@ main (int argc, char **argv) {
 
     close(fd);
     close(temp_fd);
-    free(buf);
     return 0;
 }
